Tests/LibWeb/Text/input/HTML/HTMLSelectElement-options-available-at-load.html

<!DOCTYPE html>
<script src="../include.js"></script>
<select id="select-test">
    <option>Option 1</option>
    <option>Option 2</option>
</select>
<script>
    test(() => {
        const s = document.getElementById("select-test");
        if (s.value === "Option 1" && s.selectedOptions[0].selected) {
            println("PASS (correct value)");
        }
    });
</script>
