Tests/LibWeb/Text/expected/HTML/HTMLSelectElement-options-available-at-load.txt

Option 1  PASS (correct value)
